refactor(test): cleaner tests
